🐛 Fix seek-preview with new hls.js config
The previously used hls.js config for the seek-preview had
`autoStartLoad:false` which caused errors, when using the seek preview.

There is also little to no sense in having the seek-preview not
autoStartLoading, because it is only initialized on hover and
then it should immediately start auto loading the contents of the video.